requirements.txt
aiogram==3.8.0
langchain==0.2.12
langchain-community==0.2.10
langchain-openai==0.1.20
chromadb==0.5.3
unstructured[pdf]==0.14.5
PyPDF2==3.0.1
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
tiktoken==0.7.0
httpx==0.27.0
sentence-transformers==2.7.0
torch==2.3.1    # <-- ДОБАВЛЕНО (требуется для sentence-transformers)